Module Typing.Type_consume_expr

val type_consume_expr : Typed_ast.expr -> Typed_ast.identifier list -> Typed_ast.identifier list Core.Or_error.t

Takes as arguments the list of consumed variables before this expression, and returns the list of consumed variables after this expression is abstractly interpreted.