Module Typing__Type_consume_expr

Type checks the consume operation - ensures you can't consume or access already consumed variables

val type_consume_expr : Typing.Typed_ast.expr -> Typing.Typed_ast.identifier list -> Typing.Typed_ast.identifier list Core.Or_error.t

Takes as arguments the list of consumed variables before this expression, and returns the list of consumed variables after this expression is abstractly interpreted.